# spec/openapi/drone-backend.openapi.json
{
  "components" : {
    "schemas" : {
...
      "LogFileVO" : {
        "properties" : {
          "directory" : {
            "type" : "boolean"
          },
          "lastModified" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "size" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "LogVO" : {
        "description" : "日志内容",
        "properties" : {
          "logs" : {
            "description" : "日志行列表（按时间顺序）",
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "ResultListLogFileVO" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）",
            "items" : {
              "$ref" : "#/components/schemas/LogFileVO"
            },
            "type" : "array"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultLogVO" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/LogVO",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ResultMapStringObject" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "additionalProperties" : { },
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）",
            "type" : "object"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultTaskVo" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/TaskVo",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultVoid" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "TaskVo" : {
        "properties" : {
          "acceptTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "actionHint" : {
            "type" : "string"
          },
          "completeNote" : {
            "type" : "string"
          },
          "createTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "deviceId" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "liveState" : {
            "type" : "string"
          },
          "needManualQuote" : {
            "type" : "boolean"
          },
          "orderNum" : {
            "type" : "string"
          },
          "orderStatus" : {
            "type" : "string"
          },
          "priceDetail" : {
            "type" : "string"
          },
          "referencePrice" : {
            "type" : "number"
          },
          "riderId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "riderName" : {
            "type" : "string"
          },
          "taskName" : {
            "type" : "string"
          },
          "taskNum" : {
            "type" : "string"
          },
          "taskStatus" : {
            "enum" : [ "COMPLETED", "IDLE", "IN_PROGRESS" ],
            "type" : "string"
          },
          "taskTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "taskType" : {
            "enum" : [ "AERIAL_PHOTO", "MONITORING", "SURVEY", "TRANSPORT" ],
            "type" : "string"
          },
          "totalAmount" : {
            "type" : "number"
          },
          "totalDistance" : {
            "type" : "number"
          },
          "updateTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "waypoints" : {
            "items" : {
              "$ref" : "#/components/schemas/WaypointVo"
            },
            "type" : "array"
          },
          "weight" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
      "WaypointVo" : {
        "properties" : {
          "altitude" : {
            "format" : "double",
            "type" : "number"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "latitude" : {
            "format" : "double",
            "type" : "number"
          },
          "longitude" : {
            "format" : "double",
            "type" : "number"
          },
          "orderIndex" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "Bearer Authentication" : {
        "bearerFormat" : "JWT",
        "description" : "请输入JWT Token，格式：直接输入token字符串即可",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "无人机项目web后端接口文档",
    "title" : "无人机项目接口文档",
    "version" : "1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/admin/logs/application" : {
      "get" : {
        "description" : "读取应用日志文件（application.log）的最后 N 行",
        "operationId" : "getApplicationLogs",
        "parameters" : [ {
          "description" : "读取行数，默认 100",
          "in" : "query",
          "name" : "lines",
          "required" : false,
          "schema" : {
            "default" : 100,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultLogVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "获取应用日志",
        "tags" : [ "Log API" ]
      }
    },
    "/admin/logs/error" : {
      "get" : {
        "description" : "读取错误日志文件（error.log）的最后 N 行",
        "operationId" : "getErrorLogs",
        "parameters" : [ {
          "description" : "读取行数，默认 100",
          "in" : "query",
          "name" : "lines",
          "required" : false,
          "schema" : {
            "default" : 100,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultLogVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "获取错误日志",
        "tags" : [ "Log API" ]
      }
    },
    "/admin/logs/files" : {
      "get" : {
        "description" : "列出日志目录下的文件和子目录",
        "operationId" : "listFiles_1",
        "parameters" : [ {
          "description" : "相对路径，留空表示根目录",
          "in" : "query",
          "name" : "path",
          "required" : false,
          "schema" : {
            "default" : "",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultListLogFileVO"
                }
              }
            },
            "description" : "获取成功"
          }
        },
        "summary" : "获取日志文件列表",
        "tags" : [ "Log API" ]
      }
    },
    "/admin/logs/read" : {
      "get" : {
        "description" : "读取指定日志文件的最后 N 行",
        "operationId" : "readLog",
        "parameters" : [ {
          "description" : "日志文件相对路径",
          "in" : "query",
          "name" : "file",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "读取行数，默认 200",
          "in" : "query",
          "name" : "lines",
          "required" : false,
          "schema" : {
            "default" : 200,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultLogVO"
                }
              }
            },
            "description" : "读取成功"
          }
        },
        "summary" : "读取日志文件内容",
        "tags" : [ "Log API" ]
      }
    },
...
    "/api/ws/request" : {
      "post" : {
        "description" : "设备申请建立WebSocket连接",
        "operationId" : "requestConnection",
        "parameters" : [ {
          "in" : "query",
          "name" : "deviceId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "申请成功"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "申请失败"
          }
        },
        "summary" : "申请WebSocket连接",
        "tags" : [ "APP WebSocket API" ]
      }
    },
...
    "/rider/accept" : {
      "post" : {
        "description" : "骑手接受指定任务",
        "operationId" : "acceptTask",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "query",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "接受任务",
        "tags" : [ "Rider API" ]
      }
    },
    "/rider/cancel" : {
      "post" : {
        "description" : "骑手取消已接受的任务，任务回到空闲状态",
        "operationId" : "cancelTask",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "query",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "取消接单",
        "tags" : [ "Rider API" ]
      }
    },
    "/rider/complete" : {
      "post" : {
        "description" : "骑手完成已接受的任务；note 为可选完成说明（≤500 字）",
        "operationId" : "completeTask",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "query",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "完成说明（可选，≤500 字）",
          "in" : "query",
          "name" : "note",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "完成任务",
        "tags" : [ "Rider API" ]
      }
    },
...
    "/rider/task/detail" : {
      "get" : {
        "description" : "飞手查看任意任务的详细信息（无需任务归属）",
        "operationId" : "getTaskDetail_1",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "query",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTaskVo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "任务详情",
        "tags" : [ "Rider API" ]
      }
    },
...
    "/task/detail" : {
      "get" : {
        "description" : "根据任务编号获取详细信息，包含航点列表",
        "operationId" : "getTaskDetail",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "query",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTaskVo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "获取任务详情",
        "tags" : [ "Task API" ]
      }
    },
...
    "/upload/list" : {
      "get" : {
        "description" : "支持按 orderNum 筛选，不传则查当前用户全部文件",
        "operationId" : "listFiles",
        "parameters" : [ {
          "in" : "query",
          "name" : "orderNum",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultMapStringObject"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "文件列表",
        "tags" : [ "Upload API" ]
      }
    },
...
  },
  "security" : [ {
    "Bearer Authentication" : [ ]
  } ],
  "servers" : [ {
    "description" : "相对基址：由部署环境（网关/反向代理）决定",
    "url" : "/"
  } ],
...
}